pyproject.toml
[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[project]
name = "lbrc_flask"
version = "2.0.3"
authors = [
  {name="Richard Bramley", email="rabramley@gmail.com" },
]
description = "NIHR Leicester BRC Flask Components and Theme"
readme = "README.md"
requires-python = "==3.13.*"
classifiers = [
    "Programming Language :: Python :: 3",
    "License :: OSI Approved :: MIT License",
    "Operating System :: OS Independent",
]
dependencies=[
  'alembic',
  'arrow',
  'bcrypt',
  'beautifulsoup4',
  'cairosvg',
  'celery',
  'chardet',
  'email_validator',
  'Faker',
  'Flask',
  'flask_mail',
  'flask-admin==2.0.0a4',
  'flask-security-too[fsqla,common]',
  'flask_sqlalchemy',
  'flask_wtf',
  'flask_weasyprint',
  'hatchling',
  'jsonschema',
  'ldap3',
  'markdown',
  'openpyxl',
  'pygal',
  'pymysql',
  'pytest',
  'pytest-xdist',
  'pytest-cov',
  'python-dateutil',
  'python-dotenv',
  'pytz',
  'rich',
  'setuptools',
  'WTForms',
  'WeasyPrint',
  'xlrd',
  'xlwt',
  'lxml',
]

[project.urls]
Homepage = "https://github.com/LCBRU/lbrc_flask"
Issues = "https://github.com/LCBRU/lbrc_flask/issues"

[tool.pytest.ini_options]
pythonpath = "src/"
markers = [
    'app_crsf',
]